tests/wpt/web-platform-tests/speculation-rules/prefetch/duplicate-urls.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.sub.js"></script>

<script>
  promise_test(async t => {
    assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

    let urls = Array(5).fill(getPrefetchUrlList(1)[0]);
    insertSpeculationRules({ prefetch: [{ source: 'list', urls: urls }] });
    await new Promise(resolve => t.step_timeout(resolve, 2000));

    let prefetched_count = (await Promise.all(urls.map(isUrlPrefetched))).reduce(
      (count, was_prefetched) => count + (was_prefetched ? 1 : 0), 0);

    assert_equals(prefetched_count, 1, "url should be prefetched just once.");
  }, "browser should remove duplicate urls from prefetch buffer.");
</script>
